Start breaking up store service
